refactor: removes the React.Fragments
